Add files via upload
to start the game, download all files to your computer and save them in a separate folder, you can use visual studio to read the files, hit open folder and select the folder you saved the files in. then you can hit run on the top toolbox and open it using a browser.